+++ lib/src/utils/enums.dart
@@ -1,3 +1,6 @@
 enum HttpResponseType { FULL_RESPONSE, JSON, BODY_BYTES, STRING }
 
 enum HttpMethodType { GET, POST, DELETE, PUT }
+
+/// enum for email and phone types.
+enum MaskType { auto, email, phone }